+++ packages/ui/nodes/excalidraw/src/components/ExcalidrawToolbarButton/ExcalidrawToolbarButton.tsx
@@ -23,18 +23,14 @@ export const ExcalidrawToolbarButton = ({
 
   return (
     <ToolbarButton
+      tooltip={{ content: 'Excalidraw' }}
       active={isExcalidraw}
-      onMouseDown={async (event: any) => {
-        if (!editor) return;
+      onClick={(e) => {
+        e.preventDefault();
+        e.stopPropagation();
 
-        event.preventDefault();
-        event.stopPropagation();
-
-        focusEditor(editor, editor.selection ?? editor.prevSelection!);
-
-        setTimeout(() => {
-          insertExcalidraw(editor);
-        }, 0);
+        insertExcalidraw(editor);
+        focusEditor(editor);
       }}
       {...props}
     />