+++ packages/ui/nodes/list/src/ListToolbarButton/ListToolbarButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  getPreventDefaultHandler,
+  focusEditor,
   useEventPlateId,
   usePlateEditorState,
 } from '@udecode/plate-core';
@@ -23,12 +23,13 @@ export const ListToolbarButton = ({
     <BlockToolbarButton
       active={!!res && res.list[0].type === type}
       type={type}
-      onMouseDown={
-        editor &&
-        getPreventDefaultHandler(toggleList, editor, {
-          type,
-        })
-      }
+      onClick={(e) => {
+        e.preventDefault();
+        e.stopPropagation();
+
+        toggleList(editor, { type });
+        focusEditor(editor);
+      }}
       {...props}
     />
   );