+++ packages/ui/toolbar/src/ToolbarButton/ToolbarButton.tsx
@@ -1,21 +1,26 @@
 import React from 'react';
-import Tippy, { TippyProps } from '@tippyjs/react'; // optional
+import Tippy, { TippyProps } from '@tippyjs/react';
 import clsx from 'clsx';
 import { getToolbarButtonStyles } from './ToolbarButton.styles';
 import { ToolbarButtonProps } from './ToolbarButton.types';
 
 export const ToolbarButton = (props: ToolbarButtonProps) => {
-  const { icon, tooltip, onMouseDown } = props;
-
-  const spanProps = {
-    onMouseDown,
-  };
+  const {
+    id,
+    active: _active,
+    icon,
+    tooltip,
+    className,
+    actionHandler = 'onClick',
+    onClick,
+    ...buttonProps
+  } = props;
 
   const tooltipProps: TippyProps = {
     content: '',
-    arrow: true,
-    offset: [0, 17],
-    delay: 0,
+    offset: [0, 5],
+    arrow: false,
+    delay: 500,
     duration: [200, 0],
     hideOnClick: false,
     ...tooltip,
@@ -23,15 +28,20 @@ export const ToolbarButton = (props: ToolbarButtonProps) => {
 
   const { root, active } = getToolbarButtonStyles(props);
 
+  // this can replace onClick by onMouseDown
+  buttonProps[actionHandler] = onClick;
+
   const button = (
-    <span
+    <button
       data-testid="ToolbarButton"
+      type="button"
+      aria-label={tooltipProps.content as string}
       css={root.css}
-      className={clsx(root.className, active?.className)}
-      {...spanProps}
+      className={clsx(root.className, active?.className, className)}
+      {...buttonProps}
     >
       {icon}
-    </span>
+    </button>
   );
 
   return tooltip ? <Tippy {...tooltipProps}>{button}</Tippy> : button;