+++ packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @udecode/plate-core
 
+## 19.7.0
+
+### Minor Changes
+
+- [#2225](https://github.com/udecode/plate/pull/2225) by [@TomMorane](https://github.com/TomMorane) – vendor: upgrade react-hotkeys-hook to v4
+
+### Patch Changes
+
+- [#2233](https://github.com/udecode/plate/pull/2233) by [@fimion](https://github.com/fimion) – Fixes #2230: infinite recursion when using plugin field `then`
+
+## 19.5.0
+
+### Patch Changes
+
+- [#2211](https://github.com/udecode/plate/pull/2211) by [@zbeyens](https://github.com/zbeyens) –
+  - support `slate/slate-react@0.90.0`
+  - add `isElement(n)` to `isBlock` as it has been removed by slate
+  - Fixes #2203
+  - Fixes #2197
+
 ## 19.4.4
 
 ### Patch Changes