+++ packages/editor/tabbable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @udecode/plate-tabbable
 
+## 19.7.0
+
+## 19.5.0
+
+### Patch Changes
+
+- [#2202](https://github.com/udecode/plate/pull/2202) by [@zbeyens](https://github.com/zbeyens) – Patch
+
+## 19.4.5
+
+### Patch Changes
+
+- [#2205](https://github.com/udecode/plate/pull/2205) by [@12joan](https://github.com/12joan) – Make plugin options optional
+
 ## 19.4.3
 
 ### Patch Changes