+++ packages/serializers/html/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-serializer-html
 
+## 19.7.0
+
+## 19.5.0
+
 ## 19.4.4
 
 ## 19.4.2