+++ packages/ui/nodes/alignment/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @udecode/plate-alignment-ui
 
+## 19.7.0
+
+## 19.6.0
+
+### Minor Changes
+
+- [#2212](https://github.com/udecode/plate/pull/2212) by [@TomMorane](https://github.com/TomMorane) – Toolbar buttons: add default tooltip content
+
+## 19.5.0
+
+### Minor Changes
+
+- [#2202](https://github.com/udecode/plate/pull/2202) by [@zbeyens](https://github.com/zbeyens) – Replace onMouseDown by onClick. Add aria-label.
+
 ## 19.4.4
 
 ## 19.4.2